refactor(homeView): reorganize section definitions  (#6010)
* refactor: SimilarToRecentlyViewedArtworks section

* refactor: CuratorsPicksEmerging section

* refactor: RecentlyViewedArtworks section

* refactor: AuctionLotsForYou section

* refactor: NewWorksForYou section

* refactor: NewWorksFromGalleriesYouFollow section

* refactor: RecommendedArtworks section

* refactor: ActiveBids section

* refactor: TrendingArtists section

* refactor: RecommendedArtists section

* refactor: HeroUnits section

* refactor: FeaturedFairs section

* refactor: MarketingCollections section

* refactor: ShowsForYou section

* refactor: ViewingRooms section

* refactor: LatestActivity section

* refactor: LatestAuctionResults section

* refactor: News section

* refactor: LatestArticles section

* refactor: Auctions section

* refactor: GalleriesNearYou section

* chore(analytics): add ownerType value to home view viewAll signals

Because I eagerly merged #6013, I needed to cherry-pick those changes
and thread them through to the new section files :face_palm:


---------
