+++ src/schema/v2/homeView/sections/GalleriesNearYou.ts
@@ -1,0 +1,21 @@
+import { ContextModule } from "@artsy/cohesion"
+import { HomeViewSection } from "."
+import { HomeViewSectionTypeNames } from "../HomeViewSection"
+
+export const GalleriesNearYou: HomeViewSection = {
+  id: "home-view-section-galleries-near-you",
+  type: HomeViewSectionTypeNames.HomeViewSectionGalleries,
+  contextModule: ContextModule.galleriesForYouBanner,
+  component: {
+    title: "Galleries Near You",
+    description:
+      "Follow these local galleries for updates on artists you love.",
+    backgroundImageURL: "https://files.artsy.net/images/galleries_for_you.webp",
+    behaviors: {
+      viewAll: {
+        buttonText: "Explore",
+      },
+    },
+  },
+  requiresAuthentication: false,
+}
