+++ src/schema/v2/homeView/sections/ShowsForYou.ts
@@ -1,0 +1,13 @@
+import { ContextModule } from "@artsy/cohesion"
+import { HomeViewSection } from "."
+import { HomeViewSectionTypeNames } from "../HomeViewSection"
+
+export const ShowsForYou: HomeViewSection = {
+  id: "home-view-section-shows-for-you",
+  type: HomeViewSectionTypeNames.HomeViewSectionShows,
+  contextModule: ContextModule.showsRail,
+  component: {
+    title: "Shows for You",
+  },
+  requiresAuthentication: true,
+}
