+++ src/schema/v2/homeView/zones/legacy.ts
@@ -1,28 +1,26 @@
 import { ResolverContext } from "types/graphql"
-import {
-  Auctions,
-  AuctionLotsForYou,
-  CuratorsPicksEmerging,
-  FeaturedFairs,
-  HeroUnits,
-  HomeViewSection,
-  LatestActivity,
-  LatestArticles,
-  LatestAuctionResults,
-  MarketingCollections,
-  NewWorksForYou,
-  NewWorksFromGalleriesYouFollow,
-  News,
-  RecentlyViewedArtworks,
-  RecommendedArtists,
-  ShowsForYou,
-  SimilarToRecentlyViewedArtworks,
-  TrendingArtists,
-  ViewingRooms,
-  RecommendedArtworks,
-  ActiveBids,
-  GalleriesNearYou,
-} from "schema/v2/homeView/sections"
+import { HomeViewSection } from "schema/v2/homeView/sections"
+import { GalleriesNearYou } from "../sections/GalleriesNearYou"
+import { Auctions } from "../sections/Auctions"
+import { LatestArticles } from "../sections/LatestArticles"
+import { News } from "../sections/News"
+import { LatestAuctionResults } from "../sections/LatestAuctionResults"
+import { LatestActivity } from "../sections/LatestActivity"
+import { ViewingRooms } from "../sections/ViewingRooms"
+import { ShowsForYou } from "../sections/ShowsForYou"
+import { MarketingCollections } from "../sections/MarketingCollections"
+import { FeaturedFairs } from "../sections/FeaturedFairs"
+import { HeroUnits } from "../sections/HeroUnits"
+import { RecommendedArtists } from "../sections/RecommendedArtists"
+import { TrendingArtists } from "../sections/TrendingArtists"
+import { ActiveBids } from "../sections/ActiveBids"
+import { RecommendedArtworks } from "../sections/RecommendedArtworks"
+import { NewWorksFromGalleriesYouFollow } from "../sections/NewWorksFromGalleriesYouFollow"
+import { NewWorksForYou } from "../sections/NewWorksForYou"
+import { AuctionLotsForYou } from "../sections/AuctionLotsForYou"
+import { RecentlyViewedArtworks } from "../sections/RecentlyViewedArtworks"
+import { CuratorsPicksEmerging } from "../sections/CuratorsPicksEmerging"
+import { SimilarToRecentlyViewedArtworks } from "../sections/SimilarToRecentlyViewedArtworks"
 
 const LEGACY_ZONE_SECTIONS: HomeViewSection[] = [
   LatestActivity,
